META-17: Analytics: Fix updating 3.2 database
VMS-13968: remove assert
META-204: Fix GlobalToolTip behavior
VMS-13948: Fix client crash when migrating the old settings
VMS-13894 Updated dialogs with error messages for downloading and installing states during system update.
VMS-13922: Server: ignore error on setting SO_REUSEPORT in case of this flag is not supported
META-245: Ignore object track best shot packets with a negative timestamp
META-247: Added the confidence getter to the IAttribute interface
VMS-13858: add missed log message
VMS-13986: Panoramic cameras look panoramic in thumbnails
VMS-14012: Analytics DB: protection from crash in case of DB initialization error
VMS-14012: Removed redundant #include
VMS-14005: Fix a bug (a dangling pointer) in the updatePermissions() function
META-196: Made the IHandler interface reference countable
META-196: Made the IAttribute interface reference countable
META-196: Made the IStringList interface reference countable
META-196: DeepStream plugin fixes
META-238: Don't recreate a device agent when a device status goes form Online to Recording
VMS-13937: Add missed nx_vms_utils library, needed for testcamera
VMS-13849: Fix for infinite requesting chunks from the same faulty peer
VMS-13822: Fix memory leak in the Interactive Settings component
VMS-12804: Process correctly a case when Client installs update from the already downloaded file
VMS-13491: Added more logs to ClientUpdateTool
VMS-14006: Impossible to manually add Hanwha cameras to server by ONVIF
VMS-13697: Add TextArea component to interactive settings engine
VMS-13929: Hide device-dependent analytics engines without parameters from Analytics Engines list
VMS-14019, VMS-1624: Refactored aspect ratio and geometry calculation for cameras put on a layout
VMS-11283: Use specific icons for Mac DMG installers if available
META-248: Deprecate object preview information nx.sys attributes on the client
META-248: Deprecate object preview information nx.sys attributes on the client
VMS-13929: Hide all analytics engines without settings from Analytics Engines list
VMS-13491: Improved installZip command for aplauncher
ENV-203: Support git in cmake scripts
ENV-203: Fix usage of old hg_changeset function in cmake
ENV-203: Rename mercurial.cmake to vcs_helpers.cmake
META-212: Fix potential concurrency-related crash in Analytics metadata logging
VMS-13712: Forbid to install windows server with build number lower than already installed
ENV-203: Fix cmake style
VMS-13811: Update dialog will switch to 'internet update' when update is finished
VMS-13742: Now user can select analytics storage even when there is yet no analytics
VMS-13742: Changed warnings layout in Server Settings / Storage Management tab
VMS-13742: Now inaccessible analytics storage can't be removed until another storage is selected for analytics
VMS-14015: Live button in Navigator is checked now when adding a camera without archive into the scene
VMS-13766: Set default connect timeout to 3 minutes

<==== Mercurial data ====>
Branch: vms

+++ vms/client/applauncher/src/task_server_new.cpp
@@ -1,24 +1,20 @@
-////////////////////////////////////////////////////////////
-// 14 aug 2013    Andrey Kolesnikov
-////////////////////////////////////////////////////////////
-
 #include "task_server_new.h"
 
-#ifndef _WIN32
+#if !defined(WIN32)
 #include <unistd.h>
 #endif
 
 #include <qtsinglecoreapplication.h>
 
 #include <nx/utils/log/log.h>
+#include <nx/utils/app_info.h>
 
 #include "abstract_request_processor.h"
 
+namespace applauncher {
 
-TaskServerNew::TaskServerNew( AbstractRequestProcessor* const requestProcessor )
-:
-    m_requestProcessor( requestProcessor ),
-    m_terminated( false )
+TaskServerNew::TaskServerNew(AbstractRequestProcessor* const requestProcessor):
+    m_requestProcessor(requestProcessor)
 {
 }
 
@@ -32,56 +28,59 @@ void TaskServerNew::pleaseStop()
     m_terminated = true;
 }
 
-bool TaskServerNew::listen( const QString& pipeName )
+bool TaskServerNew::listen(const QString& pipeName)
 {
-    if( isRunning() )
+    if (isRunning())
     {
         pleaseStop();
         wait();
     }
 
-    //on windows multiple m_taskServer can listen single pipe,
-        //on linux unix socket can hang after server crash (because of socket file not removed)
-    QtSingleCoreApplication* singleApp = qobject_cast<QtSingleCoreApplication*>(QCoreApplication::instance());
-    NX_ASSERT( singleApp );
-    if( singleApp->isRunning() )
+    // On Windows multiple m_taskServer can listen a single pipe, on linux unix socket can hang
+    // after server crash (because of socket file not removed).
+    const auto app = qobject_cast<QtSingleCoreApplication*>(QCoreApplication::instance());
+    if (!NX_ASSERT(app))
+        return false;
+
+    if (app->isRunning())
     {
-        NX_DEBUG(this, QString::fromLatin1("Application instance already running. Not listening to pipe"));
+        NX_DEBUG(this, "Application instance is already running. Not listening to pipe.");
         return false;
     }
 
-#ifndef _WIN32
-    //removing unix socket file in case it hanged after process crash
-    const QByteArray filePath = (QLatin1String("/tmp/")+pipeName).toLatin1();
-    unlink( filePath.constData() );
-#endif
-
-    const SystemError::ErrorCode osError = m_server.listen( pipeName );
-    if( osError != SystemError::noError )
+    if (!nx::utils::AppInfo::isWindows())
     {
-        NX_DEBUG(this, QString::fromLatin1("Failed to listen to pipe %1. %2").arg(pipeName).arg(SystemError::toString(osError)));
+        // Removing unix socket file in case it hanged after a process crash.
+        const QByteArray filePath = ("/tmp/" + pipeName).toLatin1();
+        unlink(filePath.constData());
+    }
+
+    const SystemError::ErrorCode osError = m_server.listen(pipeName);
+    if (osError != SystemError::noError)
+    {
+        NX_DEBUG(this, "Failed to listen to pipe %1. %2",
+            pipeName, SystemError::toString(osError));
         return false;
     }
 
     m_pipeName = pipeName;
 
-    NX_DEBUG(this, QString::fromLatin1("Listening pipe %1").arg(pipeName));
+    NX_DEBUG(this, "Listening pipe %1", pipeName);
 
-    start();    //launching thread
+    start();
 
     return true;
 }
 
-static const int ERROR_SLEEP_TIMEOUT_MS = 1000;
-static const int MAX_MILLIS_TO_WAIT_CLIENT_CONNECTION = 1000;
-
 void TaskServerNew::run()
 {
+    static const int kErrorSleepTimeoutMs = 1000;
+    static const int kClientConnectionTimeoutMs = 1000;
+
     while (!m_terminated)
     {
         NamedPipeSocket* clientConnection = nullptr;
-        const auto osError =
-            m_server.accept(&clientConnection, MAX_MILLIS_TO_WAIT_CLIENT_CONNECTION);
+        const auto osError = m_server.accept(&clientConnection, kClientConnectionTimeoutMs);
 
         std::unique_ptr<NamedPipeSocket> clientConnectionPtr(clientConnection);
 
@@ -90,11 +89,9 @@ void TaskServerNew::run()
             if (osError == SystemError::timedOut)
                 continue;
 
-            NX_DEBUG(this,
-                lm("Failed to listen to pipe %1. %2")
-                    .arg(m_pipeName).arg(SystemError::toString(osError)));
+            NX_DEBUG(this, "Failed to listen to pipe %1. %2", m_pipeName, osError);
 
-            msleep(ERROR_SLEEP_TIMEOUT_MS);
+            msleep(kErrorSleepTimeoutMs);
             continue;
         }
 
@@ -102,38 +99,37 @@ void TaskServerNew::run()
     }
 }
 
-static const int MAX_MSG_SIZE = 4*1024;
-
-void TaskServerNew::processNewConnection( NamedPipeSocket* clientConnection )
+void TaskServerNew::processNewConnection(NamedPipeSocket* clientConnection)
 {
-    char msgBuf[MAX_MSG_SIZE];
+    static const int kMaxMessageSize = 4 * 1024;
+
+    char buffer[kMaxMessageSize];
 
     unsigned int bytesRead = 0;
-    SystemError::ErrorCode result = clientConnection->read( msgBuf, MAX_MSG_SIZE, &bytesRead );
-    if( result != SystemError::noError )
+    SystemError::ErrorCode result = clientConnection->read(buffer, kMaxMessageSize, &bytesRead);
+    if (result != SystemError::noError)
     {
-        NX_DEBUG(this, QString::fromLatin1("Failed to receive task data. %1").arg(SystemError::toString(result)));
+        NX_DEBUG(this, "Failed to receive task data. %1", result);
         return;
     }
 
-    applauncher::api::BaseTask* task = NULL;
-    if( !applauncher::api::deserializeTask( QByteArray::fromRawData(msgBuf, bytesRead), &task ) )
+    api::BaseTask* task = nullptr;
+    if (!deserializeTask(QByteArray::fromRawData(buffer, (int) bytesRead), &task))
     {
-        NX_DEBUG(this, QString::fromLatin1("Failed to deserialize received task data"));
+        NX_DEBUG(this, "Failed to deserialize received task data");
         return;
     }
 
-    if( task->type == applauncher::api::TaskType::quit )
+    if (task->type == api::TaskType::quit)
         m_terminated = true;
 
-    //m_taskQueue->push( std::shared_ptr<applauncher::api::BaseTask>(task) );
-    applauncher::api::Response* response = NULL;
-    m_requestProcessor->processRequest(
-        std::shared_ptr<applauncher::api::BaseTask>(task),
-        &response );
+    api::Response* response = nullptr;
+    m_requestProcessor->processRequest(std::shared_ptr<api::BaseTask>(task), &response);
 
-    const QByteArray& responseMsg = response == NULL ? "ok\n\n" : response->serialize();
+    const QByteArray& responseMsg = (response == nullptr) ? "ok\n\n" : response->serialize();
     unsigned int bytesWritten = 0;
-    clientConnection->write( responseMsg.constData(), responseMsg.size(), &bytesWritten );
+    clientConnection->write(responseMsg.constData(), (unsigned) responseMsg.size(), &bytesWritten);
     clientConnection->flush();
 }
+
+} // namespace applauncher
